requirements.txt
fastapi==0.109.0
uvicorn==0.25.0
asyncio==3.4.3
azure-eventhub==5.11.5
python-dotenv==1.0.0